refactor: move log export from RuntimeEngineManager to AppDelegate
Log export is an app-level concern, not engine management.
